Attempt at migration to latest sveltekit version (#33)
* Update libs - move prerender from config to layout

* svelte-migrate routes script

* svelte-migrate routes script

* Routing Changes

* migration tasks - add data prop

* migration tasks - add data prop

* fix blog posts

* fix rss
