+++ src/lib/layout/article.svelte
@@ -1,12 +1,13 @@
 <script>
+
 	import Header from '$lib/components/layout/header.svelte';
 	import Footer from '$lib/components/layout/footer.svelte';
 	import Image from '$lib/components/base/image.svelte';
 	import { siteBaseUrl } from '$lib/meta';
 
-	export let title;
-	export let image;
-	export let excerpt;	
+	export let data;
+	let {title,image,excerpt} = data;
+
 </script>
 
 <svelte:head>
