+++ src/routes/(blog-article)/[slug]/+page.svelte
@@ -1,0 +1,7 @@
+<script>
+	export let data;
+	let {page} = data;
+
+</script>
+
+<svelte:component this={page}/>
