+++ src/routes/(waves)/404/+page.svelte
