+++ src/routes/(waves)/blog/+page.svelte
@@ -1,30 +1,13 @@
-<script context="module">
-	export async function load({ fetch }) {
-		const url = '/blog.json';
-		const res = await fetch(url);
 
-		if (res.ok) {
-			return {
-				props: {
-					posts: await res.json()
-				}
-			};
-		}
-
-		return {
-			status: res.status,
-			error: new Error(`Could not load ${url}`)
-		};
-	}
-</script>
 
 <script>
 	import BlogGrid from '$lib/components/layout/3-2-1-grid.svelte';
 	import Section from '$lib/components/layout/section.svelte';
 	import BlogPostCard from '$lib/components/base/blog-post-card.svelte';
 
-	export let posts;
-</script>
+	export let data;
+
+	 let {posts} = data;</script>
 
 <section id="blog">
 	<Section align="top" title="All posts">
