+++ src/routes/+layout.svelte
@@ -1,6 +1,8 @@
 <script>
 	import '../app.scss';
 	import { keywords } from '$lib/meta';
+	export const prerender = true;
+
 </script>
 
 <svelte:head>
