+++ src/routes/rss.xml/+server.js
@@ -7,11 +7,9 @@ export async function GET() {
 		'Cache-Control': 'max-age=0, s-maxage=3600',
 		'Content-Type': 'application/xml'
 	};
-	return {
-		headers,
-		body
-	};
+  return new Response( body , { headers });
 }
+
 
 const xml = posts => `
 <rss xmlns:dc="https://purl.org/dc/elements/1.1/" xmlns:content="https://purl.org/rss/1.0/modules/content/" xmlns:atom="https://www.w3.org/2005/Atom" version="2.0">
